refactor: move API logic to Next.js routes and update endpoints
