+++ app/api/api.ts
@@ -1,0 +1,8 @@
+import axios from 'axios';
+
+const baseURL = process.env.BACKEND_URL;
+
+export const serverApi = axios.create({
+  baseURL,
+  withCredentials: true,
+});
