+++ app/api/auth/logout/route.ts
@@ -1,14 +1,25 @@
 import { NextResponse } from 'next/server';
+import { serverApi } from '@/app/api/api';
+import { getAuthHeaders } from '@/lib/api/serverApi';
 import { cookies } from 'next/headers';
+import { AxiosError } from 'axios';
 
 export async function POST() {
   try {
-    const cookieStore = await cookies();
+    const headers = await getAuthHeaders();
 
+    await serverApi.post('/auth/logout', {}, { headers });
+
+    const cookieStore = await cookies();
     cookieStore.delete('token');
 
     return new NextResponse(null, { status: 204 });
-  } catch {
-    return NextResponse.json({ message: 'Logout error' }, { status: 500 });
+  } catch (error) {
+    const err = error as AxiosError<{ message: string }>;
+
+    return NextResponse.json(
+      { message: err.response?.data?.message || 'Logout error' },
+      { status: err.response?.status || 500 },
+    );
   }
 }
