+++ app/api/stories/popular/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { serverApi } from '@/lib/api/serverApi';
+import { serverApi } from '@/app/api/api';
 
 export async function GET() {
   try {
